requirements.txt
PyPDF2>=3.0.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
vaderSentiment>=3.3.2
transformers>=4.30.0
torch>=2.0.0
nltk>=3.8.1
python-dotenv>=1.0.0
sentence-transformers>=2.5.1
scikit-learn>=1.4.0
numpy>=1.24.0
pdfminer.six>=20221105
